update free uses gui
